# docs/developers/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/sales-orders": {
      "get": {
        "operationId": "OrdersController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "post": {
        "operationId": "OrdersController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}": {
      "get": {
        "operationId": "OrdersController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "patch": {
        "operationId": "OrdersController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/state": {
      "patch": {
        "operationId": "OrdersController_changeState",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/archive": {
      "post": {
        "operationId": "OrdersController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/unarchive": {
      "post": {
        "operationId": "OrdersController_unarchive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/lines": {
      "post": {
        "operationId": "OrdersController_addLine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderLineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/lines/{lineId}": {
      "patch": {
        "operationId": "OrdersController_updateLine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderLineDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      },
      "delete": {
        "operationId": "OrdersController_removeLine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
    "/sales-orders/{id}/post-confirmation-lines": {
      "post": {
        "operationId": "OrdersController_addPostConfirmationLine",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderLineDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "Orders"
        ]
      }
    },
...
    "/goods-received": {
      "post": {
        "operationId": "GoodsReceivedController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateGoodsReceivedDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      },
      "get": {
        "operationId": "GoodsReceivedController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
    "/goods-received/lines": {
      "get": {
        "operationId": "GoodsReceivedController_findAllLines",
        "parameters": [
          {
            "name": "purchaseOrderId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "putawayStatus",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "locationId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
    "/goods-received/{id}": {
      "get": {
        "operationId": "GoodsReceivedController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
    "/goods-received/{id}/cancel": {
      "post": {
        "operationId": "GoodsReceivedController_cancelReception",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
    "/goods-received/lines/{lineId}/resolve": {
      "post": {
        "operationId": "GoodsReceivedController_resolveAllocation",
        "parameters": [
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveAllocationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
    "/goods-received/lines/{lineId}/unresolve": {
      "post": {
        "operationId": "GoodsReceivedController_unresolveAllocation",
        "parameters": [
          {
            "name": "lineId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "GoodsReceived"
        ]
      }
    },
...
  },
  "info": {
    "title": "ModBM API",
    "description": "Core Forgeron API System endpoints",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "CreateOrderLineDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "productDescription": {
            "type": "string"
          },
          "quantity": {
            "type": "string"
          },
          "pricePerUnit": {
            "type": "string"
          },
          "discountPercentage": {
            "type": "string"
          },
          "taxCategoryId": {
            "type": "string"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "fulfillmentLocationId": {
            "type": "string"
          }
        },
        "required": [
          "quantity",
          "pricePerUnit"
        ]
      },
      "CreateOrderDto": {
        "type": "object",
        "properties": {
          "salesOrderId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "customerOrderNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "fulfillmentLocationId": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateOrderLineDto"
            }
          }
        },
        "required": [
          "salesOrderId",
          "customerId",
          "lines"
        ]
      },
      "UpdateOrderDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "customerOrderNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "fulfillmentLocationId": {
            "type": "string"
          }
        }
      },
      "UpdateOrderLineDto": {
        "type": "object",
        "properties": {
          "quantity": {
            "type": "string"
          },
          "pricePerUnit": {
            "type": "string"
          },
          "discountPercentage": {
            "type": "string"
          },
          "taxCategoryId": {
            "type": "string"
          },
          "productDescription": {
            "type": "string"
          },
          "unitOfMeasure": {
            "type": "string"
          },
          "fulfillmentLocationId": {
            "type": "string"
          }
        }
      },
...
      "CreateGoodsReceivedLineDto": {
        "type": "object",
        "properties": {
          "productId": {
            "type": "string"
          },
          "quantityReceived": {
            "type": "string"
          }
        },
        "required": [
          "productId",
          "quantityReceived"
        ]
      },
      "CreateGoodsReceivedDto": {
        "type": "object",
        "properties": {
          "goodsReceivedId": {
            "type": "string",
            "format": "uuid"
          },
          "vendorId": {
            "type": "string"
          },
          "locationId": {
            "type": "string"
          },
          "packingSlipNumber": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "lines": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CreateGoodsReceivedLineDto"
            }
          }
        },
        "required": [
          "goodsReceivedId",
          "vendorId",
          "locationId",
          "lines"
        ]
      },
      "ResolveAllocationDto": {
        "type": "object",
        "properties": {
          "purchaseOrderLineId": {
            "type": "string"
          },
          "allocatedQuantity": {
            "type": "string"
          }
        },
        "required": [
          "purchaseOrderLineId"
        ]
      },
...
    }
  }
}